Consolidate AdminInsights files
