+++ src/admin/insights/AdminAppInsightsIcon.tsx
