+++ src/admin/insights/index.ts
@@ -1,0 +1,7 @@
+export type AdminAppInsight =
+  'fork' |
+  'forkBehind' |
+  'ai' |
+  'aiRateLimiting' |
+  'photoMatting' |
+  'gridFirst';
